==> notebooks/01_load_raw_data.ipynb <==
# 01 - Load Raw Data

# import libraries
import pandas as pd
import requests
from pathlib import Path
from typing import Optional

def download_raw_data(year: int, month: int) -> pd.DataFrame:
    """
    Downloads the raw data for a given year and month from a specified URL and saves it as a parquet file.

    Args:
        year (int): The year of the data to download.
        month (int): The month of the data to download.

    Returns:
        str: The path to the downloaded parquet file.

    Raises:
        Exception: If the download fails.

    Example:
        >>> download_raw_data(2023, 1)
        '../data/raw/yellow_tripdata_2023-01.parquet'
    """
    URL_PATH = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet' 

    response = requests.get(URL_PATH)

    if response.status_code == 200:
        path = f"../data/raw/yellow_tripdata_{year}-{month:02d}.parquet"
        open(path, 'wb').write(response.content)
        return path
    else:
        raise Exception(f"Failed to download data from {URL_PATH}")


download_raw_data(year=2023, month=1)

# load raw data in parquet format using pandas
df = pd.read_parquet('../data/raw/yellow_tripdata_2023-01.parquet')
df.head(30)

df = df[['tpep_pickup_datetime', 'PULocationID']]
df.head()

# renomear as colunas par um nome mais conveniente
df.rename(
    columns={
        'tpep_pickup_datetime': 'pickup_datetime',
        'PULocationID':'pickup_location_id'
    },
    inplace=True
    )

df.head()

df.describe()

# validacao dos dados
df = df[df['pickup_datetime'] < '2023-02-01']
df = df[df['pickup_datetime'] >= '2023-01-01']

df.pickup_datetime.describe()

df.to_parquet('../data/transformed/validated_yellow_tripdata_2023_01.parquet')
