==> notebooks/02_exploratory_data_analysis.ipynb <==
# Estou pnensando na ETL
# Fluxo que eu quero executar
# RAW -> VALIDATED -> TIME SERIES DATA -> TRAINING DATA

# import libraries
import pandas as pd
from tqdm import tqdm
from typing import Optional, List
import plotly.express as px

# Load the validated raw data and check the first 20 entries
rides = pd.read_parquet('../data/transformed/validated_yellow_tripdata_2023_01.parquet')
rides.head(20)

rides['pickup_datetime'] = rides.pickup_datetime.dt.floor('H')

rides

agg = rides.groupby(['pickup_datetime', 'pickup_location_id']).size().reset_index()
agg.rename(columns={0: 'rides'}, inplace=True)

agg

# Imagine que h'a locais que n~ao tem dados para determinada hora... fill the gap


def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    """
    Add missing hourly slots to a DataFrame of ride counts, filling them with zeros.

    Parameters:
    - agg_rides (pd.DataFrame): A DataFrame containing ride counts, with columns 'pickup_hour', 'pickup_location_id', and 'rides'.

    Returns:
    - pd.DataFrame: A new DataFrame with the same columns as `agg_rides`, but with additional rows for any missing hourly slots, filled with zeros.

    Example:
    >>> rides = pd.DataFrame({'pickup_hour': ['2022-01-01 00:00:00', '2022-01-01 02:00:00'], 'pickup_location_id': [1, 1], 'rides': [10, 20]})
    >>> add_missing_slots(rides)
    pickup_hour         pickup_location_id   rides
    2022-01-01 00:00:00 1                    10
    2022-01-01 01:00:00 1                    0
    2022-01-01 02:00:00 1                    20
    """
    
    location_ids = agg_rides['pickup_location_id'].unique()
    full_range = pd.date_range(agg_rides['pickup_datetime'].min(), agg_rides['pickup_datetime'].max(), freq='H')
    output = pd.DataFrame()

    for location_id in tqdm(location_ids):
        agg_rides_i = agg_rides.loc[agg_rides['pickup_location_id'] == location_id, ['pickup_datetime', 'rides']]
        agg_rides_i.set_index('pickup_datetime', inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)
        agg_rides_i['pickup_location_id'] = location_id

        output = pd.concat([output, agg_rides_i])

    output = output.reset_index().rename(columns={'index': 'pickup_datetime'})

    return output


agg_rides_all_slots = add_missing_slots(agg)

agg_rides_all_slots.loc[agg_rides_all_slots['pickup_location_id'] == 1]

def plot_rides(
        rides: pd.DataFrame,
        locations: Optional[List[int]] = None
) -> None:
    """
    Plot ride counts over time, for one or more locations.
    """
    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides

    fig = px.line(
        rides_to_plot,
        x='pickup_datetime',
        y='rides',
        color='pickup_location_id',
        title='Rides over time'
    )

    fig.show()

plot_rides(agg_rides_all_slots, locations=[43])

agg_rides_all_slots.to_parquet('../data/transformed/ts_data_rides_2022_01.parquet')
